flux_fhir_analysis::conv

Function type_param_predicates

fn type_param_predicates<'tcx>(
    tcx: TyCtxt<'tcx>,
    item_def_id: DefId,
    param_id: DefId,
) -> impl Iterator<Item = PolyTraitPredicate<'tcx>>

This is like [TyCtxt::type_param_predicates] but computes all bounds not just the ones defining an associated item. We must compute this ourselves to resolve type-relative associated refinements, but we also use it to resolve type-relative type paths.

NOTE: [TyCtxt::type_param_predicates] is defined specifically to avoid cycles which is not a problem for us so we can use it instead of [TyCtxt::type_param_predicates].