flux_fhir_analysis

Module wf

Checks type well-formedness

Well-formedness checking assumes names are correctly bound which is guaranteed after desugaring.

Modules

errors ğŸ”’

param_usage ğŸ”’

Code to check whether refinement parameters are used in allowed positions.

sortck ğŸ”’

Structs

RefineParamVisitor ğŸ”’

Wf ğŸ”’

Functions

check_constant_expr ğŸ”’

check_flux_item ğŸ”’

check_invariants ğŸ”’

check_node ğŸ”’

visit_refine_params ğŸ”’

Type Aliases

Result ğŸ”’