flux_fhir_analysis::wf::param_usage

Struct ParamUsesChecker

struct ParamUsesChecker<'a, 'genv, 'tcx> {
    infcx: &'a InferCtxt<'genv, 'tcx>,
    xi: SnapshotMap<ParamId, ()>,
    errors: Errors<'genv>,
}

Fields
infcx: &'a InferCtxt<'genv, 'tcx>xi: SnapshotMap<ParamId, ()>

Keeps track of all refinement parameters that are used as an index such that their value is fully determined. The name xi is taken from 1, where the well-formedness judgment uses an uppercase Xi (Ξ) for a context that is similar in purpose.

This is basically a set of [fhir::ParamId] implemented with a snapshot map such that elements can be removed in batch when there’s a change in polarity.

errors: Errors<'genv>
Implementations

impl<'a, 'genv, 'tcx> ParamUsesChecker<'a, 'genv, 'tcx>

fn new(infcx: &'a InferCtxt<'genv, 'tcx>) -> Self

fn run(self, f: impl FnOnce(&mut Self)) -> Result<(), ErrorGuaranteed>

fn insert_value_determined(&mut self, expr: &Expr<'_>)

Insert params that are considered to be value determined to xi.

fn check_func_params_uses(&mut self, expr: &Expr<'_>, is_top_level_conj: bool)

Checks that refinement parameters of function sort are used in allowed positions.

fn check_params_are_value_determined(&mut self, params: &[RefineParam<'_>])

Check that Hindly parameters in params appear in a value determined position