flux_fhir_analysis::wf::sortck

Struct InferCtxt

pub(super) struct InferCtxt<'genv, 'tcx> {

    pub genv: GlobalEnv<'genv, 'tcx>,
    pub owner: FluxOwnerId,
    pub wfckresults: WfckResults,
    sort_unification_table: InPlaceUnificationTable<SortVid>,
    num_unification_table: InPlaceUnificationTable<NumVid>,
    bv_size_unification_table: InPlaceUnificationTable<BvSizeVid>,
    params: FxHashMap<ParamId, (RefineParam<'genv>, Sort)>,
    sort_of_bty: FxHashMap<FhirId, Sort>,
    path_args: UnordMap<FhirId, GenericArgs>,
    sort_of_alias_reft: FxHashMap<FhirId, FuncSort>,
    sort_of_literal: NodeMap<'genv, Sort>,
    sort_of_bin_op: NodeMap<'genv, Sort>,
    sort_args_of_app: NodeMap<'genv, List<SortArg>>,

}

Fields
genv: GlobalEnv<'genv, 'tcx>owner: FluxOwnerIdwfckresults: WfckResultssort_unification_table: InPlaceUnificationTable<SortVid>num_unification_table: InPlaceUnificationTable<NumVid>bv_size_unification_table: InPlaceUnificationTable<BvSizeVid>params: FxHashMap<ParamId, (RefineParam<'genv>, Sort)>sort_of_bty: FxHashMap<FhirId, Sort>path_args: UnordMap<FhirId, GenericArgs>sort_of_alias_reft: FxHashMap<FhirId, FuncSort>sort_of_literal: NodeMap<'genv, Sort>sort_of_bin_op: NodeMap<'genv, Sort>sort_args_of_app: NodeMap<'genv, List<SortArg>>
Implementations

impl<'genv, 'tcx> InferCtxt<'genv, 'tcx>

pub(super) fn new(genv: GlobalEnv<'genv, 'tcx>, owner: FluxOwnerId) -> Self

fn check_abs( &mut self, expr: &Expr<'genv>, params: &[RefineParam<'_>], body: &Expr<'genv>, expected: &Sort, ) -> Result<(), ErrorGuaranteed>

fn check_field_exprs( &mut self, expr_span: Span, sort_def: &AdtSortDef, sort_args: &[Sort], field_exprs: &[FieldExpr<'genv>], spread: &Option<&Spread<'genv>>, expected: &Sort, ) -> Result<(), ErrorGuaranteed>

fn check_constructor( &mut self, expr: &Expr<'_>, field_exprs: &[FieldExpr<'genv>], spread: &Option<&Spread<'genv>>, expected: &Sort, ) -> Result<(), ErrorGuaranteed>

fn check_record( &mut self, arg: &Expr<'genv>, flds: &[Expr<'genv>], expected: &Sort, ) -> Result<(), ErrorGuaranteed>

pub(super) fn check_expr( &mut self, expr: &Expr<'genv>, expected: &Sort, ) -> Result<(), ErrorGuaranteed>

pub(super) fn check_loc( &mut self, loc: &PathExpr<'_>, ) -> Result<(), ErrorGuaranteed>

fn synth_lit(&mut self, lit: Lit, expr: &Expr<'genv>) -> Sort

fn synth_prim_app( &mut self, op: &BinOp, e1: &Expr<'genv>, e2: &Expr<'genv>, span: Span, ) -> Result<Sort, ErrorGuaranteed>

fn synth_expr(&mut self, expr: &Expr<'genv>) -> Result<Sort, ErrorGuaranteed>

fn synth_path(&mut self, path: &PathExpr<'_>) -> Result<Sort, ErrorGuaranteed>

fn check_integral( &mut self, op: BinOp, sort: &Sort, span: Span, ) -> Result<(), ErrorGuaranteed>

fn synth_binary_op( &mut self, expr: &Expr<'genv>, op: BinOp, e1: &Expr<'genv>, e2: &Expr<'genv>, ) -> Result<Sort, ErrorGuaranteed>

fn synth_unary_op( &mut self, op: UnOp, e: &Expr<'genv>, ) -> Result<Sort, ErrorGuaranteed>

fn synth_app( &mut self, fsort: FuncSort, args: &[Expr<'genv>], span: Span, ) -> Result<Sort, ErrorGuaranteed>

fn instantiate_func_sort( &mut self, app_expr: &Expr<'genv>, fsort: PolyFuncSort, ) -> FuncSort

pub(crate) fn insert_sort_for_bty(&mut self, fhir_id: FhirId, sort: Sort)

pub(crate) fn sort_of_bty(&self, fhir_id: FhirId) -> Sort

pub(crate) fn insert_path_args(&mut self, fhir_id: FhirId, args: GenericArgs)

pub(crate) fn path_args(&self, fhir_id: FhirId) -> GenericArgs

pub(crate) fn insert_sort_for_alias_reft( &mut self, fhir_id: FhirId, fsort: FuncSort, )

fn sort_of_alias_reft(&self, fhir_id: FhirId) -> FuncSort

fn normalize_projection_sort( genv: GlobalEnv<'_, '_>, owner: FluxOwnerId, sort: Sort, ) -> Sort

pub(crate) fn normalize_sorts(&mut self) -> QueryResult

impl<'genv> InferCtxt<'genv, '_>

pub(super) fn declare_param(&mut self, param: RefineParam<'genv>, sort: Sort)

fn is_coercible(&mut self, sort1: &Sort, sort2: &Sort, fhir_id: FhirId) -> bool

Whether a value of sort1 can be automatically coerced to a value of sort2. A value of an [rty::SortCtor::Adt] sort with a single field of sort s can be coerced to a value of sort s and vice versa, i.e., we can automatically project the field out of the record or inject a value into a record.

fn is_coercible_from_func( &mut self, sort: &Sort, fhir_id: FhirId, ) -> Option<PolyFuncSort>

fn is_coercible_to_func( &mut self, sort: &Sort, fhir_id: FhirId, ) -> Option<PolyFuncSort>

fn try_equate(&mut self, sort1: &Sort, sort2: &Sort) -> Option<Sort>

fn try_equate_inner(&mut self, sort1: &Sort, sort2: &Sort) -> Option<Sort>

fn try_equate_bv_sizes( &mut self, size1: BvSize, size2: BvSize, ) -> Option<BvSize>

fn equate(&mut self, sort1: &Sort, sort2: &Sort) -> Sort

pub(crate) fn next_sort_var(&mut self) -> Sort

fn next_num_var(&mut self) -> Sort

pub(crate) fn next_sort_vid(&mut self) -> SortVid

fn next_num_vid(&mut self) -> NumVid

fn next_bv_size_var(&mut self) -> BvSize

fn next_bv_size_vid(&mut self) -> BvSizeVid

fn ensure_resolved_path( &mut self, path: &PathExpr<'_>, ) -> Result<Sort, ErrorGuaranteed>

fn is_single_field_struct(&mut self, sort: &Sort) -> Option<(DefId, Sort)>

pub(crate) fn into_results(self) -> Result<WfckResults, ErrorGuaranteed>

pub(crate) fn infer_mode(&self, id: ParamId) -> InferMode

pub(crate) fn param_sort(&self, id: ParamId) -> Sort

fn shallow_resolve(&mut self, sort: &Sort) -> Sort

fn resolve_vars_if_possible(&mut self, sort: &Sort) -> Sort

pub(crate) fn fully_resolve(&mut self, sort: &Sort) -> Result<Sort, ()>

impl InferCtxt<'_, '_>

fn emit_sort_mismatch( &mut self, span: Span, expected: &Sort, found: &Sort, ) -> ErrorGuaranteed

fn emit_field_not_found(&mut self, sort: &Sort, field: Ident) -> ErrorGuaranteed

fn emit_err<'b>(&'b self, err: impl Diagnostic<'b>) -> ErrorGuaranteed