flux_fhir_analysis::wf
Struct Wf

struct Wf<'a, 'genv, 'tcx> {
    infcx: &'a mut InferCtxt<'genv, 'tcx>,
    errors: Errors<'genv>,
    next_type_index: u32,
    next_region_index: u32,
    next_const_index: u32,
}

Fields
infcx: &'a mut InferCtxt<'genv, 'tcx>errors: Errors<'genv>next_type_index: u32next_region_index: u32next_const_index: u32
Implementations

impl<'a, 'genv, 'tcx> Wf<'a, 'genv, 'tcx>

fn new(infcx: &'a mut InferCtxt<'genv, 'tcx>) -> Self

fn check_node(&mut self, node: &OwnerNode<'genv>)

fn insert_params( &mut self, params: &[RefineParam<'_>], ) -> Result<(), ErrorGuaranteed>

Push a layer of binders. We assume all names are fresh so we don’t care about shadowing

fn insert_params_for_node( &mut self, node: &OwnerNode<'_>, ) -> Result<(), ErrorGuaranteed>

Initializes the inference context with all refinement parameters in node

fn init_infcx(&mut self, node: &OwnerNode<'genv>) -> QueryResult

To check for well-formedness we need to know the sort of base types. For example, to check if the type i32[e] is well formed, we need to know that the sort of i32 is int so we can check the expression e against it. Computing the sort from a base type is subtle and hard to do in fhir so we must do it in rty. However, to convert from fhir to rty we need elaborated information from sort checking which we do in fhir.

To break this circularity, we do conversion in two phases. In the first phase, we do conversion without elaborated information. This results in types in rty with incorrect refinements but with the right shape to compute their sorts. We use these sorts for sort checking and then do conversion again with the elaborated information.

This function initializes the inference context by running the first phase of conversion and collecting the sort of all base types.

fn check_output_locs(&mut self, fn_decl: &FnDecl<'_>)