Module flux_fhir_analysis::conv

Conversion from types in [fhir] to types in [rty]

Conversion assumes well-formedness and will panic if type are not well-formed. Among other things, well-formedness implies:

Names are bound correctly.

Refinement parameters appear in allowed positions. This is particularly important for refinement predicates, aka abstract refinements, since the syntax in [rty] has syntactic restrictions on predicates.

Refinements are well-sorted.

Modules

errors ğŸ”’

Structs

ConvCtxt

Env ğŸ”’

Layer ğŸ”’

LookupResult ğŸ”’

Enums

Entry ğŸ”’

LookupResultKind ğŸ”’

Functions

adt_def_for_enum ğŸ”’

adt_def_for_struct ğŸ”’

conv_defn ğŸ”’

conv_fn_sig ğŸ”’

conv_func_decl

conv_func_sort ğŸ”’

conv_generic_predicates ğŸ”’

conv_generics ğŸ”’

conv_invariants ğŸ”’

conv_lit ğŸ”’

conv_opaque_ty ğŸ”’

conv_qualifier

conv_refine_param ğŸ”’

conv_sort

conv_sort_ctor ğŸ”’

conv_sorts ğŸ”’

conv_ty ğŸ”’

def_id_to_param_index ğŸ”’

def_id_to_param_ty ğŸ”’

expand_type_alias ğŸ”’