Module flux_fhir_analysis::conv

Conversion from types in [fhir] to types in [rty]

Conversion assumes well-formedness and will panic if type are not well-formed. Among other things, well-formedness implies:

Names are bound correctly.

Refinement parameters appear in allowed positions. This is particularly important for refinement predicates, aka abstract refinements, since the syntax in [rty] has syntactic restrictions on predicates.

Refinements are well-sorted.

Modules

errors ğŸ”’

Structs

ConvCtxt

Env ğŸ”’

Layer ğŸ”’

LookupResult ğŸ”’

Enums

Entry ğŸ”’

LayerKind ğŸ”’

LookupResultKind ğŸ”’

Functions

bug_on_infer_sort ğŸ”’

conv_adt_sort_def ğŸ”’

conv_assoc_reft_def ğŸ”’

conv_defn ğŸ”’

conv_fn_decl ğŸ”’

conv_func_decl

conv_func_sort ğŸ”’

conv_generic_param_kind ğŸ”’

conv_generic_predicates ğŸ”’

conv_generics ğŸ”’

conv_invariants ğŸ”’

conv_lit ğŸ”’

conv_opaque_ty ğŸ”’

conv_poly_func_sort ğŸ”’

conv_qualifier ğŸ”’

conv_refine_param ğŸ”’

conv_refinement_generics ğŸ”’

conv_sort ğŸ”’

conv_sort_path ğŸ”’

conv_sorts ğŸ”’

conv_ty ğŸ”’

conv_un_op ğŸ”’

expand_type_alias ğŸ”’

resolve_param_sort ğŸ”’