flux_fhir_analysis::conv

Trait WfckResultsProvider

pub trait WfckResultsProvider: Sized {
    // Required methods
    fn bin_op_sort(&self, fhir_id: FhirId) -> Sort;
    fn coercions_for(&self, fhir_id: FhirId) -> &[Coercion];
    fn field_proj(&self, fhir_id: FhirId) -> FieldProj;
    fn record_ctor(&self, fhir_id: FhirId) -> DefId;
    fn param_sort(&self, param_id: ParamId) -> Sort;
    fn node_sort(&self, fhir_id: FhirId) -> Sort;
    fn node_sort_args(&self, fhir_id: FhirId) -> List<SortArg>;
}

An interface to the information elaborated during sort checking. We mock these results in the first conversion phase during sort checking.

Required Methods

fn bin_op_sort(&self, fhir_id: FhirId) -> Sort

fn coercions_for(&self, fhir_id: FhirId) -> &[Coercion]

fn field_proj(&self, fhir_id: FhirId) -> FieldProj

fn record_ctor(&self, fhir_id: FhirId) -> DefId

fn param_sort(&self, param_id: ParamId) -> Sort

fn node_sort(&self, fhir_id: FhirId) -> Sort

fn node_sort_args(&self, fhir_id: FhirId) -> List<SortArg>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.