Module flux_fhir_analysis::wf

Checks type well-formedness

Well-formedness checking assumes names are correctly bound which is guaranteed after desugaring.

Modules

errors ğŸ”’

param_usage ğŸ”’

Code to check whether refinement parameters are used in allowed positions.

sortck ğŸ”’

Structs

Wf ğŸ”’

Functions

check_flux_item ğŸ”’

check_fn_quals ğŸ”’

check_node ğŸ”’

insert_params ğŸ”’

Initializes the inference context with all parameters required to check node

resolve_params ğŸ”’

Check that all params with fhir::Sort::Infer have a sort inferred and save it in the WfckResults

visit_refine_params ğŸ”’

Type Aliases

Result ğŸ”’