Module flux_fhir_analysis::wf

Checks type well-formedness

Well-formedness checking assumes names are correctly bound which is guaranteed after desugaring.

Modules

errors ğŸ”’

param_usage ğŸ”’

Code to check whether refinement parameters are used in allowed positions.

sortck ğŸ”’

Structs

ParamSortResolver ğŸ”’

Check that all params with fhir::Sort::Infer have a sort inferred and elaborate it

Wf ğŸ”’

Functions

check_enum_def ğŸ”’

check_fn_decl ğŸ”’

check_fn_quals ğŸ”’

check_impl ğŸ”’

check_opaque_ty ğŸ”’

check_qualifier ğŸ”’

check_spec_func ğŸ”’

check_struct_def ğŸ”’

check_ty_alias ğŸ”’

Type Aliases

Result ğŸ”’