Module flux_fhir_analysis::wf

Checks type well-formedness

Well-formedness checking assumes names are correctly bound which is guaranteed after desugaring.

Modules

errors ğŸ”’

sortck ğŸ”’

Structs

Wf ğŸ”’

XiCtxt ğŸ”’

Keeps track of all refinement parameters that are used as an index such that their value is fully determined. The context is called Xi because in the paper Focusing on Liquid Refinement Typing, the well-formedness judgment uses an uppercase Xi (Î) for a context that is similar in purpose.

Functions

check_defn ğŸ”’

check_enum_def ğŸ”’

check_fn_quals ğŸ”’

check_fn_sig ğŸ”’

check_opaque_ty ğŸ”’

check_qualifier ğŸ”’

check_struct_def ğŸ”’

check_ty_alias ğŸ”’

check_type ğŸ”’