Struct flux_fhir_analysis::wf::sortck::InferCtxt

pub(super) struct InferCtxt<'a, 'tcx> {
    pub genv: &'a GlobalEnv<'a, 'tcx>,
    sorts: UnordMap<Name, Sort>,
    unification_table: InPlaceUnificationTable<SortVid>,
    wfckresults: WfckResults,
    eq_vids: HashSet<SortVid>,
}

Fields
genv: &'a GlobalEnv<'a, 'tcx>sorts: UnordMap<Name, Sort>unification_table: InPlaceUnificationTable<SortVid>wfckresults: WfckResultseq_vids: HashSet<SortVid>

sort variables that can only be instantiated to sorts that support equality (i.e. non FuncSort)

Implementations

impl<'a, 'tcx> InferCtxt<'a, 'tcx>

pub(super) fn new(genv: &'a GlobalEnv<'a, 'tcx>, owner: FluxOwnerId) -> Self

pub(super) fn check_refine_arg( &mut self, arg: &RefineArg, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_abs( &mut self, params: &Vec<RefineParam>, body: &Expr, span: &Span, fhir_id: &FhirId, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_record( &mut self, def_id: DefId, sort_args: &[Sort], args: &[RefineArg], span: Span ) -> Result<(), ErrorGuaranteed>

pub(super) fn check_expr( &mut self, expr: &Expr, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_binary_op( &mut self, expr: &Expr, op: BinOp, e1: &Expr, e2: &Expr, expected: &Sort ) -> Result<(), ErrorGuaranteed>

pub(super) fn check_loc(&mut self, loc: Ident) -> Result<(), ErrorGuaranteed>

fn synth_expr(&mut self, expr: &Expr) -> Result<Sort, ErrorGuaranteed>

fn synth_binary_op( &mut self, expr: &Expr, op: BinOp, e1: &Expr, e2: &Expr ) -> Result<Sort, ErrorGuaranteed>

fn synth_unary_op( &mut self, op: UnOp, e: &Expr ) -> Result<Sort, ErrorGuaranteed>

fn synth_app( &mut self, func: &Func, args: &[Expr], span: Span ) -> Result<Sort, ErrorGuaranteed>

fn synth_func(&mut self, func: &Func) -> Result<FuncSort, ErrorGuaranteed>

fn instantiate_func_sort(&mut self, fsort: PolyFuncSort) -> FuncSort

impl<'a> InferCtxt<'a, '_>

pub(super) fn push_layer<'b>( &mut self, params: impl IntoIterator<Item = &'b RefineParam> )

Push a layer of binders. We assume all names are fresh so we don’t care about shadowing

fn is_coercible(&mut self, sort1: &Sort, sort2: &Sort, fhir_id: FhirId) -> bool

Whether a value of sort1 can be automatically coerced to a value of sort2. A value of a Record sort with a single field of sort s can be coerced to a value of sort s and vice versa, i.e., we can automatically project the field out of the record or inject a value into a record.

fn is_coercible_to_numeric( &mut self, sort: &Sort, fhir_id: FhirId ) -> Option<Sort>

fn is_coercible_to_func( &mut self, sort: &Sort, fhir_id: FhirId ) -> Option<PolyFuncSort>

fn try_equate(&mut self, sort1: &Sort, sort2: &Sort) -> Option<Sort>

fn next_sort_vid(&mut self) -> SortVid

fn next_eq_sort_vid(&mut self) -> SortVid

fn is_eq_sort_vid(&self, vid: SortVid) -> bool

pub(crate) fn resolve_params_sorts( &mut self, params: &[RefineParam] ) -> Result<(), ErrorGuaranteed>

fn resolve_param(&mut self, param: &RefineParam) -> Option<Sort>

fn ensure_resolved_var(&mut self, var: Ident) -> Result<Sort, ErrorGuaranteed>

fn resolve_sort(&mut self, sort: &Sort) -> Option<Sort>

fn is_numeric(&mut self, sort: &Sort) -> bool

fn is_bool(&mut self, sort: &Sort) -> bool

fn is_int(&mut self, sort: &Sort) -> bool

fn is_func(&mut self, sort: &Sort) -> Option<PolyFuncSort>

fn is_single_field_record(&mut self, sort: &Sort) -> Option<Sort>

fn has_equality(&mut self, sort: &Sort) -> bool

pub(crate) fn into_results(self) -> WfckResults

impl InferCtxt<'_, '_>

fn emit_sort_mismatch( &mut self, span: Span, expected: &Sort, found: &Sort ) -> ErrorGuaranteed

fn emit_field_not_found( &mut self, sort: &Sort, field: SurfaceIdent ) -> ErrorGuaranteed

fn emit_err<'b>(&'b self, err: impl IntoDiagnostic<'b>) -> ErrorGuaranteed