Struct flux_fhir_analysis::wf::sortck::InferCtxt

pub(super) struct InferCtxt<'genv, 'tcx> {
    pub genv: GlobalEnv<'genv, 'tcx>,
    pub params: UnordMap<ParamId, (Sort, ParamKind)>,
    pub(super) sort_unification_table: InPlaceUnificationTable<SortVid>,
    num_unification_table: InPlaceUnificationTable<NumVid>,
    pub wfckresults: WfckResults,
}

Fields
genv: GlobalEnv<'genv, 'tcx>params: UnordMap<ParamId, (Sort, ParamKind)>sort_unification_table: InPlaceUnificationTable<SortVid>num_unification_table: InPlaceUnificationTable<NumVid>wfckresults: WfckResults
Implementations

impl<'genv, 'tcx> InferCtxt<'genv, 'tcx>

pub(super) fn new(genv: GlobalEnv<'genv, 'tcx>, owner: FluxOwnerId) -> Self

pub(super) fn check_refine_arg( &mut self, arg: &RefineArg<'_>, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_abs( &mut self, arg: &RefineArg<'_>, params: &[RefineParam<'_>], body: &Expr<'_>, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_record( &mut self, arg: &RefineArg<'_>, flds: &[RefineArg<'_>], expected: &Sort ) -> Result<(), ErrorGuaranteed>

pub(super) fn check_expr( &mut self, expr: &Expr<'_>, expected: &Sort ) -> Result<(), ErrorGuaranteed>

pub(super) fn check_loc( &mut self, loc: &PathExpr<'_> ) -> Result<(), ErrorGuaranteed>

fn synth_expr(&mut self, expr: &Expr<'_>) -> Result<Sort, ErrorGuaranteed>

fn synth_var(&mut self, path: &PathExpr<'_>) -> Sort

fn synth_binary_op( &mut self, expr: &Expr<'_>, op: BinOp, e1: &Expr<'_>, e2: &Expr<'_> ) -> Result<Sort, ErrorGuaranteed>

fn synth_unary_op( &mut self, op: UnOp, e: &Expr<'_> ) -> Result<Sort, ErrorGuaranteed>

fn synth_app( &mut self, func: &PathExpr<'_>, args: &[Expr<'_>], span: Span ) -> Result<Sort, ErrorGuaranteed>

fn synth_alias_reft_app( &mut self, alias: &AliasReft<'_>, args: &[Expr<'_>], span: Span ) -> Result<Sort, ErrorGuaranteed>

fn synth_func( &mut self, func: &PathExpr<'_> ) -> Result<FuncSort, ErrorGuaranteed>

fn instantiate_func_sort(&mut self, fsort: PolyFuncSort) -> FuncSort

impl<'genv> InferCtxt<'genv, '_>

pub(super) fn insert_params( &mut self, params: &[RefineParam<'_>] ) -> Result<(), ErrorGuaranteed>

Push a layer of binders. We assume all names are fresh so we don’t care about shadowing

pub(super) fn insert_param(&mut self, id: ParamId, sort: Sort, kind: ParamKind)

fn is_coercible(&mut self, sort1: &Sort, sort2: &Sort, fhir_id: FhirId) -> bool

Whether a value of sort1 can be automatically coerced to a value of sort2. A value of an rty::SortCtor::Adt sort with a single field of sort s can be coerced to a value of sort s and vice versa, i.e., we can automatically project the field out of the record or inject a value into a record.

fn is_coercible_from_func( &mut self, sort: &Sort, fhir_id: FhirId ) -> Option<PolyFuncSort>

fn is_coercible_to_func( &mut self, sort: &Sort, fhir_id: FhirId ) -> Option<PolyFuncSort>

fn try_equate(&mut self, sort1: &Sort, sort2: &Sort) -> Option<Sort>

fn try_equate_inner(&mut self, sort1: &Sort, sort2: &Sort) -> Option<Sort>

fn equate(&mut self, sort1: &Sort, sort2: &Sort) -> Sort

pub(crate) fn next_sort_var(&mut self) -> Sort

fn next_num_var(&mut self) -> Sort

fn next_sort_vid(&mut self) -> SortVid

fn next_num_vid(&mut self) -> NumVid

pub(crate) fn resolve_param_sort( &mut self, param: &RefineParam<'_> ) -> Result<(), ErrorGuaranteed>

fn ensure_resolved_var( &mut self, path: &PathExpr<'_> ) -> Result<Sort, ErrorGuaranteed>

fn is_single_field_record(&mut self, sort: &Sort) -> Option<(DefId, Sort)>

pub(crate) fn into_results(self) -> WfckResults

pub(crate) fn infer_mode(&self, id: ParamId) -> InferMode

pub(crate) fn param_sort(&self, id: ParamId) -> Sort

fn shallow_resolve(&mut self, sort: &Sort) -> Sort

fn resolve_vars_if_possible(&mut self, sort: &Sort) -> Sort

pub(crate) fn fully_resolve(&mut self, sort: &Sort) -> Result<Sort, ()>

impl InferCtxt<'_, '_>

fn emit_sort_mismatch( &mut self, span: Span, expected: &Sort, found: &Sort ) -> ErrorGuaranteed

fn emit_field_not_found(&mut self, sort: &Sort, field: Ident) -> ErrorGuaranteed

fn emit_err<'b>(&'b self, err: impl Diagnostic<'b>) -> ErrorGuaranteed