Struct flux_fhir_analysis::wf::Wf

struct Wf<'genv, 'tcx> {
    genv: GlobalEnv<'genv, 'tcx>,
    xi: XiCtxt,
}

Fields
genv: GlobalEnv<'genv, 'tcx>xi: XiCtxt
Implementations

impl<'genv, 'tcx> Wf<'genv, 'tcx>

fn new(genv: GlobalEnv<'genv, 'tcx>) -> Self

fn check_params_are_determined( &mut self, infcx: &InferCtxt<'_, '_>, params: &[RefineParam<'_>] ) -> Result<(), ErrorGuaranteed>

fn check_generic_predicates( &mut self, infcx: &mut InferCtxt<'_, '_>, predicates: &[WhereBoundPredicate<'_>] ) -> Result<(), ErrorGuaranteed>

fn check_generic_bound( &mut self, infcx: &mut InferCtxt<'_, '_>, bound: &GenericBound<'_> ) -> Result<(), ErrorGuaranteed>

fn check_opaque_ty( &mut self, infcx: &mut InferCtxt<'_, '_>, opaque_ty: &OpaqueTy<'_> ) -> Result<(), ErrorGuaranteed>

fn check_variant( &mut self, infcx: &mut InferCtxt<'_, '_>, variant: &VariantDef<'_> ) -> Result<(), ErrorGuaranteed>

fn check_fn_output( &mut self, infcx: &mut InferCtxt<'_, '_>, fn_output: &FnOutput<'_> ) -> Result<(), ErrorGuaranteed>

fn check_output_locs(&self, fn_decl: &FnDecl<'_>) -> Result<(), ErrorGuaranteed>

fn check_constraint( &mut self, infcx: &mut InferCtxt<'_, '_>, constr: &Constraint<'_> ) -> Result<(), ErrorGuaranteed>

fn check_type( &mut self, infcx: &mut InferCtxt<'_, '_>, ty: &Ty<'_> ) -> Result<(), ErrorGuaranteed>

fn check_generic_predicate( &mut self, infcx: &mut InferCtxt<'_, '_>, predicate: &WhereBoundPredicate<'_> ) -> Result<(), ErrorGuaranteed>

fn check_generic_args( &mut self, infcx: &mut InferCtxt<'_, '_>, args: &[GenericArg<'_>] ) -> Result<(), ErrorGuaranteed>

fn check_generic_arg( &mut self, infcx: &mut InferCtxt<'_, '_>, arg: &GenericArg<'_> ) -> Result<(), ErrorGuaranteed>

fn check_type_bindings( &mut self, infcx: &mut InferCtxt<'_, '_>, bindings: &[TypeBinding<'_>] ) -> Result<(), ErrorGuaranteed>

fn check_base_ty( &mut self, infcx: &mut InferCtxt<'_, '_>, bty: &BaseTy<'_> ) -> Result<(), ErrorGuaranteed>

fn check_path( &mut self, infcx: &mut InferCtxt<'_, '_>, path: &Path<'_> ) -> Result<(), ErrorGuaranteed>

fn check_refine_arg( &mut self, infcx: &mut InferCtxt<'_, '_>, arg: &RefineArg<'_>, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_expr_as_pred( &mut self, infcx: &mut InferCtxt<'_, '_>, expr: &Expr<'_> ) -> Result<(), ErrorGuaranteed>

fn check_param_uses_refine_arg( &mut self, infcx: &mut InferCtxt<'_, '_>, arg: &RefineArg<'_> ) -> Result<(), ErrorGuaranteed>

Checks that refinement parameters of function sort are used in allowed positions.

fn check_param_uses_expr( &self, infcx: &mut InferCtxt<'_, '_>, expr: &Expr<'_>, is_top_level_conj: bool ) -> Result<(), ErrorGuaranteed>

Checks that refinement parameters of function sort are used in allowed positions.

fn sort_of_bty(&self, bty: &BaseTy<'_>) -> Sort

fn emit_err<'b, R>( &'b self, err: impl IntoDiagnostic<'b> ) -> Result<R, ErrorGuaranteed>