Struct flux_fhir_analysis::wf::Wf

struct Wf<'a, 'tcx> {
    genv: &'a GlobalEnv<'a, 'tcx>,
    modes: UnordMap<Name, InferMode>,
    xi: XiCtxt,
}

Fields
genv: &'a GlobalEnv<'a, 'tcx>modes: UnordMap<Name, InferMode>xi: XiCtxt
Implementations

impl<'a, 'tcx> Wf<'a, 'tcx>

fn new(genv: &'a GlobalEnv<'a, 'tcx>) -> Self

fn insert_refine_params( &mut self, infcx: &mut InferCtxt<'_, '_>, params: &[RefineParam] )

fn check_params_are_determined( &mut self, infcx: &InferCtxt<'_, '_>, params: &[RefineParam] ) -> Result<(), ErrorGuaranteed>

fn check_generic_predicates( &mut self, infcx: &mut InferCtxt<'_, '_>, predicates: &GenericPredicates ) -> Result<(), ErrorGuaranteed>

fn check_generic_bound( &mut self, infcx: &mut InferCtxt<'_, '_>, bound: &GenericBound ) -> Result<(), ErrorGuaranteed>

fn check_opaque_ty( &mut self, infcx: &mut InferCtxt<'_, '_>, opaque_ty: &OpaqueTy ) -> Result<(), ErrorGuaranteed>

fn check_variant( &mut self, infcx: &mut InferCtxt<'_, '_>, variant: &VariantDef ) -> Result<(), ErrorGuaranteed>

fn check_fn_output( &mut self, env: &mut InferCtxt<'_, '_>, fn_output: &FnOutput ) -> Result<(), ErrorGuaranteed>

fn check_output_locs(&self, fn_sig: &FnSig) -> Result<(), ErrorGuaranteed>

fn check_constraint( &mut self, infcx: &mut InferCtxt<'_, '_>, constr: &Constraint ) -> Result<(), ErrorGuaranteed>

fn check_type( &mut self, infcx: &mut InferCtxt<'_, '_>, ty: &Ty ) -> Result<(), ErrorGuaranteed>

fn check_generic_predicate( &mut self, infcx: &mut InferCtxt<'_, '_>, predicate: &WhereBoundPredicate ) -> Result<(), ErrorGuaranteed>

fn check_ty_is_base(&self, ty: &Ty) -> Result<(), ErrorGuaranteed>

fn check_generic_args_kinds( &self, def_id: DefId, args: &[GenericArg] ) -> Result<(), ErrorGuaranteed>

fn check_generic_args( &mut self, infcx: &mut InferCtxt<'_, '_>, def_id: DefId, args: &[GenericArg] ) -> Result<(), ErrorGuaranteed>

fn check_generic_arg( &mut self, infcx: &mut InferCtxt<'_, '_>, arg: &GenericArg ) -> Result<(), ErrorGuaranteed>

fn check_type_bindings( &mut self, infcx: &mut InferCtxt<'_, '_>, bindings: &[TypeBinding] ) -> Result<(), ErrorGuaranteed>

fn check_base_ty( &mut self, infcx: &mut InferCtxt<'_, '_>, bty: &BaseTy ) -> Result<(), ErrorGuaranteed>

fn check_path( &mut self, infcx: &mut InferCtxt<'_, '_>, path: &Path ) -> Result<(), ErrorGuaranteed>

fn check_refine_arg( &mut self, infcx: &mut InferCtxt<'_, '_>, arg: &RefineArg, expected: &Sort ) -> Result<(), ErrorGuaranteed>

fn check_pred( &mut self, infcx: &mut InferCtxt<'_, '_>, expr: &Expr ) -> Result<(), ErrorGuaranteed>

fn check_param_uses_refine_arg( &mut self, infcx: &InferCtxt<'_, '_>, arg: &RefineArg ) -> Result<(), ErrorGuaranteed>

Checks that refinement parameters of function sort are used in allowed positions.

fn check_param_uses_expr( &self, infcx: &InferCtxt<'_, '_>, expr: &Expr, is_top_level_conj: bool ) -> Result<(), ErrorGuaranteed>

Checks that refinement parameters of function sort are used in allowed positions.

fn infer_mode(&self, infcx: &InferCtxt<'_, '_>, name: Name) -> InferMode

fn sort_of_bty(&self, bty: &BaseTy) -> Sort

fn emit_err<'b, R>( &'b self, err: impl IntoDiagnostic<'b> ) -> Result<R, ErrorGuaranteed>